requirements.txt
pycryptodome~=3.23.0
requests~=2.32.4